requirements.txt
pandas==2.2.1
mysql-connector-python==8.3.0